+++ src/Make_morph.mak
@@ -49,6 +49,7 @@ SRC =	arabic.c						\
 	hardcopy.c						\
 	hashtab.c						\
 	indent.c						\
+	insexpand.c						\
 	json.c							\
 	list.c							\
 	main.c							\
