+++ src/Make_sas.mak
@@ -114,6 +114,7 @@ SRC = \
 	hardcopy.c \
 	hashtab.c \
 	indent.c \
+	insexpand.c \
 	json.c \
 	list.c \
 	main.c \
@@ -174,6 +175,7 @@ OBJ = \
 	hardcopy.o \
 	hashtab.o \
 	indent.o \
+	insexpand.o \
 	json.o \
 	list.o \
 	main.o \
@@ -234,6 +236,7 @@ PRO = \
 	proto/hardcopy.pro \
 	proto/hashtab.pro \
 	proto/indent.pro \
+	proto/insexpand.pro \
 	proto/json.pro \
 	proto/list.pro \
 	proto/main.pro \
@@ -373,6 +376,8 @@ hashtab.o:		hashtab.c
 proto/hashtab.pro:	hashtab.c
 indent.o:		indent.c
 proto/indent.pro:	indent.c
+insexpand.o:		insexpand.c
+proto/insexpand.pro:	insexpand.c
 json.o:			json.c
 proto/json.pro:		json.c
 list.o:			list.c
