+++ src/evalfunc.c
@@ -2603,7 +2603,7 @@ f_complete_check(typval_T *argvars UNUSED, typval_T *rettv)
 
     RedrawingDisabled = 0;
     ins_compl_check_keys(0, TRUE);
-    rettv->vval.v_number = compl_interrupted;
+    rettv->vval.v_number = ins_compl_interrupted();
     RedrawingDisabled = saved;
 }
