+++ src/misc2.c
@@ -1130,6 +1130,7 @@ free_all_mem(void)
     free_search_patterns();
     free_old_sub();
     free_last_insert();
+    free_insexpand_stuff();
     free_prev_shellcmd();
     free_regexp_stuff();
     free_tag_stuff();
