+++ src/proto.h
@@ -89,6 +89,9 @@ extern int _stricoll(char *a, char *b);
 # include "hardcopy.pro"
 # include "hashtab.pro"
 # include "indent.pro"
+# ifdef FEAT_INS_EXPAND
+# include "insexpand.pro"
+# endif
 # include "json.pro"
 # include "list.pro"
 # include "blob.pro"
