+++ src/tag.c
@@ -1591,7 +1591,7 @@ find_tags(
 #ifdef FEAT_INS_EXPAND
 	    if ((flags & TAG_INS_COMP))	/* Double brackets for gcc */
 		ins_compl_check_keys(30, FALSE);
-	    if (got_int || compl_interrupted)
+	    if (got_int || ins_compl_interrupted())
 #else
 	    if (got_int)
 #endif
